+++ packages/data-context/test/unit/sources/WizardDataSource.spec.ts
@@ -126,22 +126,6 @@ describe('packagesToInstall', () => {
     expect(actual).to.eq(`npm install -D next react react-dom`)
   })
 
-  it('nuxtjs-vue2-unconfigured', async () => {
-    const ctx = createTestDataContext()
-
-    const projectPath = await scaffoldMigrationProject('nuxtjs-vue2-unconfigured')
-
-    ctx.update((coreData) => {
-      coreData.currentProject = projectPath
-      coreData.wizard.chosenFramework = findFramework('nuxtjs')
-      coreData.wizard.chosenBundler = findBundler('webpack')
-    })
-
-    const actual = await ctx.wizard.installDependenciesCommand()
-
-    expect(actual).to.eq('npm install -D nuxt@2 vue@2')
-  })
-
   it('framework and bundler are undefined', async () => {
     const ctx = createTestDataContext()
     const projectPath = await scaffoldMigrationProject('pristine-with-e2e-testing')