+++ packages/frontend-shared/src/utils/icons.ts
@@ -1,7 +1,6 @@
 import LogoWebpack from '../assets/logos/webpack.svg'
 import LogoVite from '../assets/logos/vite.svg'
 import LogoNext from '../assets/logos/nextjs.svg'
-import LogoNuxt from '../assets/logos/nuxt.svg'
 import LogoVue from '../assets/logos/vue.svg'
 import LogoReact from '../assets/logos/react.svg'
 import LogoAngular from '../assets/logos/angular.svg'
@@ -15,7 +14,6 @@ export const FrameworkBundlerLogos = {
   vueclivue2: LogoVue,
   vueclivue3: LogoVue,
   nextjs: LogoNext,
-  nuxtjs: LogoNuxt,
   react: LogoReact,
   reactscripts: LogoReact,
   angular: LogoAngular,